Fixed app crashed at the first login
